Add files via upload
Excel document comprising three worksheets that, respectively, contain the parameter sets that define the nutrient limited, space limited and bistable virtual tumour populations. The parameters listed are the oxygen consumption rates, q<sub>1 and q<sub>3, the vascular volume, V<sub>0, and the steady state tumour volume and oxygen concentration, which we use as initial conditions when simulating the model.